Prepare package for boilerplate import
Applying this commit will:
- Remove bin folder
The bin folder was used for passing the start-storybook function to the package importing the toolbox, a better alternative (github-pages) was implemented instead.
- Move Button folder to src/components from src/components/demo-button
- Add index.js
This file will serve as the main export location for components for easy import in projects using the toolbox.
- Update package.json
Updated the build-storybook commands, which will now have to be ran before creating a merge request (this will be added to the README). Updated the homepage metadata and added the files parameter for package export.